# Cache parsed message templates